pyproject.toml
[project]
name = "roundtable"
description = "AI Roundtable App"
authors = [{name = "Armand McQueen", email = "armandmcqueen@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.8.2",
    "rich>=13.8.0",
    "typer>=0.12.5",
    "llmlib-amq==0.0.6",
]
dynamic = ["version"]

[tool.flit.module]
name = "roundtable"

[tool.flit.sdist]
include = ["LICENSE"]

[tool.uv]
dev-dependencies = [
    "flit>=3.9.0",
    "mypy>=1.11.2",
    "pytest>=8.3.2",
    "ruff>=0.6.3",
]

[tool.uv.sources]
roundtable = { workspace = true }




